# Conflicts:
#	calendar_scheduler/tests_calendar_scheduler/test_cancel_lessons_view.py
#	code_tutors/urls.py
#	request_handler/forms.py
#	request_handler/models.py
#	request_handler/templates/view_requests.html

+++ invoicer/tests_invoicer/test_generate_invoice.py
@@ -1,19 +1,21 @@
 import os
 
-from django.http import HttpResponse
-from django.test import TestCase, override_settings
 from django.conf import settings
-from invoicer.models import get_latest_id_number
-from user_system.models import User, KnowledgeArea, Day
-from user_system.fixtures import create_test_users as create_fixtures
-from invoicer.helpers.generate_invoice_id import generate_invoice_id
-from request_handler.models import Request, Venue
+from django.http import HttpResponse
 from django.shortcuts import reverse
+from django.test import TestCase, override_settings
+
+from invoicer.helpers.generate_invoice_id import generate_invoice_id
+from invoicer.models import get_latest_id_number
+from request_handler.models import Request, Venue
+from user_system.fixtures import create_test_users as create_fixtures
+from user_system.models import Day, KnowledgeArea, User
+
 
 class TestGenerateInvoice(TestCase):
 
     def setUp(self):
-        create_fixtures.create_test_user()
+        create_fixtures.create_test_users()
         self.student = User.objects.get(user_type='Student')
         self.tutor = User.objects.get(user_type='Tutor')
         self.admin = User.objects.get(user_type='Admin')
@@ -31,39 +33,33 @@ class TestGenerateInvoice(TestCase):
             day=Day.objects.get(day='Monday'),
         )
 
+    def tearDown(self):
+        if os.path.exists(self.path):
+            os.remove(self.path)
+        super().tearDown()
+
     def test_generate_invoice_id(self):
-        self.assertEqual(generate_invoice_id(self.student, get_latest_id_number(student=self.student)), str(self.invoice_id))
+        self.assertEqual(generate_invoice_id(self.student, get_latest_id_number(student=self.student)),
+                         str(self.invoice_id))
 
     @override_settings(USE_AWS_S3=False)
     def test_generate_invoice_local_store(self):
-        self.assertions_for_local_invoice(self.generate_invoice())
-        os.remove(self.path)
+        self._assertions_for_local_invoice(generate_invoice(self.client, self.admin, self.request.id))
 
     @override_settings(USE_AWS_S3=False)
     def test_you_cannot_generate_invoices_if_they_already_exist(self):
-        self.assertions_for_local_invoice(self.generate_invoice())
+        self._assertions_for_local_invoice(generate_invoice(self.client, self.admin, self.request.id))
         self.client.force_login(self.admin)
-        response = self.client.get(reverse('generate_invoice', kwargs={"request_id": self.request.id}))
+        response = self.client.get(reverse('generate_invoice', kwargs={"tutoring_request_id": self.request.id}))
         self.assertEqual(response.status_code, 204)
-        os.remove(self.path)
 
-    @override_settings(USE_AWS_S3=True)
-    def test_generate_invoice(self):
-        from code_tutors.aws import s3
-        self.generate_invoice()
-        key = f'invoices/pdfs/{self.invoice_id}.pdf'
-        url = s3.generate_access_url(key=key)
-        self.assertIsNotNone(url)
-        self.assertTrue(isinstance(url, str))
-        s3._delete(key)
-
-    #----TEST HELPERS-----#
-    def generate_invoice(self) -> HttpResponse:
-        self.client.force_login(self.admin)
-        return self.client.get(reverse("generate_invoice", kwargs={"request_id": self.request.id}))
-
-    def assertions_for_local_invoice(self, response: HttpResponse) -> None:
+    def _assertions_for_local_invoice(self, response: HttpResponse) -> None:
         with open(self.path, 'rb') as file:
             self.assertIsNotNone(file)
         self.assertEqual(response.status_code, 201)
         self.assertEqual(response.content, b"Invoice generated successfully!")
+
+
+def generate_invoice(client, admin, request_id) -> HttpResponse:
+    client.force_login(admin)
+    return client.get(reverse("generate_invoice", kwargs={"tutoring_request_id": request_id}))
